adapted logic for searching the positions of NE in the original text

+++ enrichment-solr/src/main/java/eu/europeana/enrichment/solr/service/SolrEntityPositionsService.java
@@ -80,7 +80,7 @@ public interface SolrEntityPositionsService {
 		 * @return 
 		 * @throws SolrNamedEntityServiceException
 		 */
-		public double findTermPositionsInStory(String storyId, String term, int startAfterOffset) throws SolrNamedEntityServiceException;
+		public int findTermPositionsInStory(String storyId, String term, int startAfterOffset) throws SolrNamedEntityServiceException;
 
 
 }
